backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
flask-migrate==4.0.5
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0

# AI/ML Libraries
transformers==4.33.2
torch==2.0.1
tensorflow==2.13.0
scikit-learn==1.3.0
opencv-python==4.8.0.76
pillow==10.0.0
numpy==1.24.3
pandas==2.0.3

# NLP Libraries
nltk==3.8.1
spacy==3.6.1
textblob==0.17.1

# OCR
pytesseract==0.3.10
easyocr==1.7.0

# QR Code Processing
pyzbar==0.1.9
qrcode==7.4.2

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.12.0

# Blockchain
web3==6.9.0
eth-account==0.9.0

# Database
psycopg2-binary==2.9.7
pymongo==4.5.0
redis==4.6.0

# Background Tasks
celery==5.3.1

# Image Processing
deepface==0.0.79
face-recognition==1.3.0

# Audio Processing
librosa==0.10.1
soundfile==0.12.1

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
dateutils==0.6.12

# Testing
pytest==7.4.2
pytest-cov==4.1.0

# Development
black==23.7.0
flake8==6.0.0